final i think. It works?

+++ lib/libft/ft_strrchr.c
@@ -6,7 +6,7 @@
 /*   By: xazuaje- <xazuaje-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/14 23:06:03 by xazuaje-          #+#    #+#             */
-/*   Updated: 2023/09/15 19:30:53 by xazuaje-         ###   ########.fr       */
+/*   Updated: 2024/01/15 02:20:43 by xazuaje-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strrchr(const char *s, int c)
 {
-	int	i;
+	size_t	i;
 
 	i = ft_strlen(s);
 	while (i >= 0)
